+++ Tests/Functional/Service/SingleViewLinkBuilderTest.php
@@ -28,6 +28,8 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
 
     private DummyConfiguration $configuration;
 
+    private EventMapper $eventMapper;
+
     private SingleViewLinkBuilder $subject;
 
     protected function setUp(): void
@@ -39,6 +41,8 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
         $this->testingFramework = $this->get(TestingFramework::class);
         $this->testingFramework->createFakeFrontEnd(1);
         $this->configuration = new DummyConfiguration();
+
+        $this->eventMapper = $this->get(MapperRegistry::class)->getByClassName(EventMapper::class);
 
         $this->subject = new SingleViewLinkBuilder($this->configuration);
     }
@@ -56,7 +60,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
      */
     public function createRelativeUrlForEventWithoutAnyConfigurationReturnsEmptyString(): void
     {
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(1);
+        $event = $this->eventMapper->find(1);
 
         $result = $this->subject->createRelativeUrlForEvent($event);
 
@@ -69,7 +73,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
     public function createRelativeUrlForEventWithConfiguredSingleViewPageReturnsLinkToItWithEventUid(): void
     {
         $this->configuration->setAsInteger('detailPID', self::DEFAULT_SINGLE_VIEW_PAGE_UID);
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(1);
+        $event = $this->eventMapper->find(1);
 
         $result = $this->subject->createRelativeUrlForEvent($event);
 
@@ -82,7 +86,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
     public function createRelativeUrlForEventWithExternalSingleViewPageReturnsLinkToExternalUrl(): void
     {
         $this->configuration->setAsInteger('detailPID', self::DEFAULT_SINGLE_VIEW_PAGE_UID);
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(2);
+        $event = $this->eventMapper->find(2);
 
         $result = $this->subject->createRelativeUrlForEvent($event);
 
@@ -95,7 +99,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
     public function createRelativeUrlWithEventTypeAndCategoryWithoutSingleViewPageUsesPageFromConfiguration(): void
     {
         $this->configuration->setAsInteger('detailPID', self::DEFAULT_SINGLE_VIEW_PAGE_UID);
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(3);
+        $event = $this->eventMapper->find(3);
 
         $result = $this->subject->createRelativeUrlForEvent($event);
 
@@ -109,7 +113,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
     public function createRelativeUrlWithEventTypeWithSingleViewPageUsesPageFromEventType(): void
     {
         $this->configuration->setAsInteger('detailPID', self::DEFAULT_SINGLE_VIEW_PAGE_UID);
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(4);
+        $event = $this->eventMapper->find(4);
 
         $result = $this->subject->createRelativeUrlForEvent($event);
 
@@ -122,7 +126,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
     public function createRelativeUrlWithCategoryWithSingleViewPageUsesPageFromCategory(): void
     {
         $this->configuration->setAsInteger('detailPID', self::DEFAULT_SINGLE_VIEW_PAGE_UID);
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(5);
+        $event = $this->eventMapper->find(5);
 
         $result = $this->subject->createRelativeUrlForEvent($event);
 
@@ -135,7 +139,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
     public function createRelativeUrlWithCategoryAndEventTypeWithSingleViewPageUsesPageFromEventType(): void
     {
         $this->configuration->setAsInteger('detailPID', self::DEFAULT_SINGLE_VIEW_PAGE_UID);
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(6);
+        $event = $this->eventMapper->find(6);
 
         $result = $this->subject->createRelativeUrlForEvent($event);
 
@@ -148,7 +152,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
     public function createRelativeUrlWithIndividualUrlAndEventTypeWithSingleViewPageUsesIndividualUrl(): void
     {
         $this->configuration->setAsInteger('detailPID', self::DEFAULT_SINGLE_VIEW_PAGE_UID);
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(7);
+        $event = $this->eventMapper->find(7);
 
         $result = $this->subject->createRelativeUrlForEvent($event);
 
@@ -160,7 +164,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
      */
     public function createAbsoluteUrlForEventWithoutAnyConfigurationReturnsBaseUrl(): void
     {
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(1);
+        $event = $this->eventMapper->find(1);
 
         $result = $this->subject->createAbsoluteUrlForEvent($event);
 
@@ -173,7 +177,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
     public function createAbsoluteUrlWithConfiguredSingleViewPageReturnsLinkToItWithEventUid(): void
     {
         $this->configuration->setAsInteger('detailPID', self::DEFAULT_SINGLE_VIEW_PAGE_UID);
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(1);
+        $event = $this->eventMapper->find(1);
 
         $result = $this->subject->createAbsoluteUrlForEvent($event);
 
@@ -187,7 +191,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
     public function createAbsoluteUrlWithExternalSingleViewPageReturnsLinkToExternalUrl(): void
     {
         $this->configuration->setAsInteger('detailPID', self::DEFAULT_SINGLE_VIEW_PAGE_UID);
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(2);
+        $event = $this->eventMapper->find(2);
 
         $result = $this->subject->createAbsoluteUrlForEvent($event);
 
@@ -200,7 +204,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
     public function createAbsoluteUrlWithEventTypeAndCategoryWithoutSingleViewPageUsesPageFromConfiguration(): void
     {
         $this->configuration->setAsInteger('detailPID', self::DEFAULT_SINGLE_VIEW_PAGE_UID);
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(3);
+        $event = $this->eventMapper->find(3);
 
         $result = $this->subject->createAbsoluteUrlForEvent($event);
 
@@ -214,7 +218,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
     public function createAbsoluteUrlWithEventTypeWithSingleViewPageUsesPageFromEventType(): void
     {
         $this->configuration->setAsInteger('detailPID', self::DEFAULT_SINGLE_VIEW_PAGE_UID);
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(4);
+        $event = $this->eventMapper->find(4);
 
         $result = $this->subject->createAbsoluteUrlForEvent($event);
 
@@ -228,7 +232,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
     public function createAbsoluteUrlWithCategoryWithSingleViewPageUsesPageFromCategory(): void
     {
         $this->configuration->setAsInteger('detailPID', self::DEFAULT_SINGLE_VIEW_PAGE_UID);
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(5);
+        $event = $this->eventMapper->find(5);
 
         $result = $this->subject->createAbsoluteUrlForEvent($event);
 
@@ -242,7 +246,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
     public function createAbsoluteUrlWithCategoryAndEventTypeWithSingleViewPageUsesPageFromEventType(): void
     {
         $this->configuration->setAsInteger('detailPID', self::DEFAULT_SINGLE_VIEW_PAGE_UID);
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(6);
+        $event = $this->eventMapper->find(6);
 
         $result = $this->subject->createAbsoluteUrlForEvent($event);
 
@@ -256,7 +260,7 @@ final class SingleViewLinkBuilderTest extends FunctionalTestCase
     public function createAbsoluteUrlWithIndividualUrlAndEventTypeWithSingleViewPageUsesIndividualUrl(): void
     {
         $this->configuration->setAsInteger('detailPID', self::DEFAULT_SINGLE_VIEW_PAGE_UID);
-        $event = MapperRegistry::getInstance()->getByClassName(EventMapper::class)->find(7);
+        $event = $this->eventMapper->find(7);
 
         $result = $this->subject->createAbsoluteUrlForEvent($event);