+++ Tests/LegacyFunctional/Mapper/OrganizerMapperTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace OliverKlee\Seminars\Tests\LegacyFunctional\Mapper;
 
+use OliverKlee\Oelib\Mapper\MapperRegistry;
 use OliverKlee\Oelib\Testing\TestingFramework;
 use OliverKlee\Seminars\Mapper\OrganizerMapper;
 use OliverKlee\Seminars\Model\Organizer;
@@ -30,11 +31,12 @@ final class OrganizerMapperTest extends FunctionalTestCase
 
         $this->testingFramework = $this->get(TestingFramework::class);
 
-        $this->subject = new OrganizerMapper();
+        $this->subject = $this->get(MapperRegistry::class)->getByClassName(OrganizerMapper::class);
     }
 
     protected function tearDown(): void
     {
+        MapperRegistry::purgeInstance();
         $this->testingFramework->cleanUpWithoutDatabase();
 
         parent::tearDown();