+++ CHANGELOG.md
@@ -235,7 +235,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Fixed
 
-- Avoid using the deprecated `MapperRegistry::getInstance()` (#5006)
+- Avoid using the deprecated `MapperRegistry::getInstance()` (#5006, #5008)
 - Avoid the deprecated `->getTypoLink()`
   (#4895, #4903, #4904, #4905, #4906, #4907)
 - Avoid accessing the deprecated flash message severity constants (#4878, #4884)